minor bug: Fixed Layout error in timeline-slice (simple circle).
In the former version, the ID was used to define the position on the circle. However, the ID may be higher than the number of vertices, for it should be provided continuously, resulting in an error. Note: One should always *fix* the layout with pajek manually. For details, see the manual of PajekToSvgAnim.